NCSD-3163: Attempt to fix resource dependencies
The initial deploy of this template always failed with a resource not found error.
This appeared to be caused by the metric exception alert not being able to find the app insights instance, despite having a dependency to it.
We suspect it's because it's setting up multiple alerts simultaneously, hence the dependency update.

I also noticed that the function app incorrectly depended on the app insights instance, when it should depend on the app service plan.

This change fixes those two issues